+++ test/test_module/future_div.py
@@ -1,8 +1,10 @@
 from __future__ import division
 import torch
 
+
 def div_int_future():
     return 1 / 2
+
 
 def div_float_future():
     return 3.14 / 0.125
