+++ test/test_module/no_future_div.py
@@ -1,7 +1,9 @@
 import torch
 
+
 def div_int_nofuture():
     return 1 / 2
+
 
 def div_float_nofuture():
     return 3.14 / 0.125
